timeDateAgent/pyproject.toml
[project]
name = "timedateagent"
version = "0.1.0"
description = "Time/Date Agent with MCP integration for time and date operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.0",
    "httpx>=0.25.0",
    "uvicorn>=0.23.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "mcp>=1.0.0",
    "pytz>=2024.1",
]

license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0"
]

